feat: Add book_id column to banners table
This commit adds a new column, book_id, to the banners table in the database. The column is of type uuid and is set to reference the id column of the books table. The column is set to not allow null values and has a cascade delete constraint.

The book_id column is also added to the NewBannerDto and UpdateBannerDto classes in the banners module. The column is validated using the IsUUID decorator with version 4.

The KyselyBannerEntity interface in the banners module is also updated to include the book_id property.

These changes are necessary to support the association of banners with specific books.